Add optional EFS VPC and subnet_id variables
This commit adds variables to enable overriding what VPC and subnet that
EFS runs in. If you don’t provide them, it default back to the 
`var.vpc_id` and `var.subnet_ids` values.

During testing of this module, we found something interesting. We were
deploying EFS to the backing services VPC, which is a different VPC to
the k8s cluster. Our pods were unable to resolve the DNS endpoint of the
EFS cluster, despite there being VPC peering between the two, with DNS
lookups between them enabled. AWS documentation [1] states that in this
scenario, you have a few options:

1) Use the EFS IP address directly (no thanks)
2) Create a DNS entry in Route53 CNAME’ing to the EFS DNS endpoint in 
   a private hosted zone

The underlying EFS module does already create a Route53 DNS entry CNAME
to the EFS DNS endpoint, but it isn’t in a private zone. The pod could
not resolve the Route53 DNS. Deploying EFS into the _same_ VPC as the 
k8s nodes worked a treat and finally the pods were able to resolve and 
mount EFS volumes.


[1] https://docs.aws.amazon.com/efs/latest/ug/manage-fs-access-vpc-peering.html